Refactor HelpContextIdProvider and WizardHelper
Moved HelpContextIdProvider and WizardHelper to solve class duplication
problem. Had to change import for several classes and function
visibility of WizardHelper
