+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/internal/wizard/project/AbstractProjectWizard.java
@@ -7,12 +7,14 @@
  *
  * Contributors:
  *     Simon Scholz <simon.scholz@vogella.com> - Bug 473545
+ *     Sebastian Kuzniarz (Diebold Nixdorf Inc.) - refactored HelpContextIdProvider
  */
 
 package org.eclipse.buildship.ui.internal.wizard.project;
 
 import org.eclipse.buildship.core.internal.GradlePluginsRuntimeException;
 import org.eclipse.buildship.ui.internal.UiPlugin;
+import org.eclipse.buildship.ui.internal.wizard.HelpContextIdProvider;
 import org.eclipse.core.runtime.preferences.ConfigurationScope;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.jface.wizard.Wizard;
